Issue #618 Add support for analytic urls
